[Matrix]: use size_t for indexing in NUMA dense matrix.

+++ matrix/NUMA_dense_matrix.cpp
@@ -87,22 +87,22 @@ NUMA_row_tall_matrix_store::NUMA_row_tall_matrix_store(size_t nrow, size_t ncol,
 				local_lens[node_id] * ncol * get_entry_size(), node_id);
 }
 
-char *NUMA_row_tall_matrix_store::get_row(off_t row_idx)
+char *NUMA_row_tall_matrix_store::get_row(size_t row_idx)
 {
 	auto phy_loc = mapper.map2physical(row_idx);
 	return data[phy_loc.first].get_raw()
 		+ phy_loc.second * get_num_cols() * get_entry_size();
 }
 
-const char *NUMA_row_tall_matrix_store::get_row(off_t row_idx) const
+const char *NUMA_row_tall_matrix_store::get_row(size_t row_idx) const
 {
 	auto phy_loc = mapper.map2physical(row_idx);
 	return data[phy_loc.first].get_raw()
 		+ phy_loc.second * get_num_cols() * get_entry_size();
 }
 
-const char *NUMA_row_tall_matrix_store::get_rows(off_t row_start,
-		off_t row_end) const
+const char *NUMA_row_tall_matrix_store::get_rows(size_t row_start,
+		size_t row_end) const
 {
 	if (mapper.get_logical_range_id(row_start)
 			!= mapper.get_logical_range_id(row_end - 1)) {
@@ -113,7 +113,7 @@ const char *NUMA_row_tall_matrix_store::get_rows(off_t row_start,
 	return get_row(row_start);
 }
 
-char *NUMA_row_tall_matrix_store::get_rows(off_t row_start, off_t row_end)
+char *NUMA_row_tall_matrix_store::get_rows(size_t row_start, size_t row_end)
 {
 	if (mapper.get_logical_range_id(row_start)
 			!= mapper.get_logical_range_id(row_end - 1)) {
